+++ src/titiler/core/titiler/core/dependencies.py
@@ -425,9 +425,29 @@ class DatasetParams(DefaultDependency):
             self.unscale = bool(self.unscale)
 
 
+RescaleType = List[Tuple[float, float]]
+
+
 @dataclass
 class ImageRenderingParams(DefaultDependency):
     """Image Rendering options."""
+
+    rescale: Annotated[
+        Optional[List[str]],
+        Query(
+            title="Min/Max data Rescaling",
+            description="comma (',') delimited Min,Max range. Can set multiple time for multiple bands.",
+            examples=["0,2000", "0,1000", "0,10000"],  # band 1  # band 2  # band 3
+        ),
+    ] = None
+
+    color_formula: Annotated[
+        Optional[str],
+        Query(
+            title="Color Formula",
+            description="rio-color formula (info: https://github.com/mapbox/rio-color)",
+        ),
+    ] = None
 
     add_mask: Annotated[
         Optional[bool],
@@ -437,8 +457,23 @@ class ImageRenderingParams(DefaultDependency):
         ),
     ] = None
 
+    def __post_init__(self):
+        """Post Init."""
+        if self.rescale:
+            rescale_array = []
+            for r in self.rescale:
+                parsed = tuple(
+                    map(
+                        float,
+                        r.replace(" ", "").replace("[", "").replace("]", "").split(","),
+                    )
+                )
+                assert (
+                    len(parsed) == 2
+                ), f"Invalid rescale values: {self.rescale}, should be of form ['min,max', 'min,max'] or [[min,max], [min, max]]"
+                rescale_array.append(parsed)
 
-RescaleType = List[Tuple[float, float]]
+            self.rescale: RescaleType = rescale_array
 
 
 def RescalingParams(
@@ -452,6 +487,11 @@ def RescalingParams(
     ] = None,
 ) -> Optional[RescaleType]:
     """Min/Max data Rescaling"""
+    warnings.warn(
+        "RescalingParams is deprecated and set to be removed in 0.20",
+        DeprecationWarning,
+        stacklevel=1,
+    )
     if rescale:
         rescale_array = []
         for r in rescale:
@@ -646,6 +686,11 @@ def ColorFormulaParams(
     ] = None,
 ) -> Optional[str]:
     """ColorFormula Parameter."""
+    warnings.warn(
+        "ColorFormulaParams is deprecated and set to be removed in 0.20",
+        DeprecationWarning,
+        stacklevel=1,
+    )
     return color_formula